+++ app/src/main/java/id/ac/umn/kevinsorensen/bengkelonline/views/user/HomeUser.kt
@@ -1,15 +1,25 @@
 package id.ac.umn.kevinsorensen.bengkelonline.views.user
 
 import android.annotation.SuppressLint
+import android.content.Context
 import android.content.Intent
+import android.content.pm.PackageManager
 import android.os.Bundle
+import android.os.Looper
 import android.util.Log
+import android.widget.Toast
 import androidx.activity.ComponentActivity
+import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.compose.setContent
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -31,20 +41,34 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
+import androidx.core.content.ContextCompat
 import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
+import com.google.android.gms.location.FusedLocationProviderClient
+import com.google.android.gms.location.LocationCallback
+import com.google.android.gms.location.LocationRequest
+import com.google.android.gms.location.LocationResult
+import com.google.android.gms.location.LocationServices
+import com.google.android.gms.location.Priority
+import com.google.android.gms.maps.MapsInitializer
 import com.google.android.gms.maps.model.CameraPosition
+import com.google.android.gms.maps.model.LatLng
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
+import com.google.maps.android.compose.CameraPositionState
+import com.google.maps.android.compose.GoogleMap
+import com.google.maps.android.compose.Marker
+import com.google.maps.android.compose.MarkerState
 import com.google.maps.android.compose.rememberCameraPositionState
 import id.ac.umn.kevinsorensen.bengkelonline.R
 import id.ac.umn.kevinsorensen.bengkelonline.api.ProductController
@@ -67,6 +91,50 @@ sealed class BottomNavItem (
 }
 
 class HomeUser : ComponentActivity() {
+    private var permissions = arrayOf(
+        android.Manifest.permission.ACCESS_COARSE_LOCATION,
+        android.Manifest.permission.ACCESS_FINE_LOCATION,
+        android.Manifest.permission.ACCESS_BACKGROUND_LOCATION,
+        android.Manifest.permission.INTERNET
+    )
+
+    private lateinit var fusedLocationClient: FusedLocationProviderClient
+    private lateinit var locationCallback: LocationCallback
+    private var locationRequired: Boolean = false
+
+    override fun onResume() {
+        super.onResume()
+        if (locationRequired) {
+            startLocationUpdates()
+        }
+    }
+
+    override fun onPause() {
+        super.onPause()
+        locationCallback?.let {
+            fusedLocationClient?.removeLocationUpdates(it)
+        }
+    }
+
+    @SuppressLint("MissingPermission")
+    private fun startLocationUpdates() {
+        locationCallback?.let {
+            val locationRequest = LocationRequest.Builder(
+                Priority.PRIORITY_HIGH_ACCURACY, 100
+            )
+                .setWaitForAccurateLocation(false)
+                .setMinUpdateIntervalMillis(3000)
+                .setMaxUpdateDelayMillis(100)
+                .build()
+
+            fusedLocationClient?.requestLocationUpdates(
+                locationRequest,
+                it,
+                Looper.getMainLooper()
+            )
+        }
+    }
+
     @OptIn(ExperimentalMaterial3Api::class)
     @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -77,8 +145,42 @@ class HomeUser : ComponentActivity() {
         val userName = intent.getStringExtra("username")?: "user";
         val profileUrl = intent.getStringExtra("profileUrl");
 
+        MapsInitializer.initialize(this, MapsInitializer.Renderer.LATEST){
+
+        }
+
+        fusedLocationClient = LocationServices.getFusedLocationProviderClient(this)
+
         setContent {
 
+            var currentLocation by remember {
+                mutableStateOf(com.google.android.gms.maps.model.LatLng(0.toDouble(), 0.toDouble()))
+            }
+
+            val cameraPosition = rememberCameraPositionState{
+                position = CameraPosition.fromLatLngZoom(
+                    currentLocation, 10f
+                )
+            }
+
+            var cameraPositionState by remember {
+                mutableStateOf(cameraPosition)
+            }
+
+            locationCallback = object: LocationCallback() {
+                override fun onLocationResult(p0: LocationResult) {
+                    super.onLocationResult(p0)
+                    for(location in p0.locations) {
+                        currentLocation = LatLng(location.latitude, location.longitude)
+
+                        cameraPositionState = CameraPositionState(
+                            position = CameraPosition.fromLatLngZoom(
+                                currentLocation, 10f
+                            )
+                        )
+                    }
+                }
+            }
 
             var products by remember { mutableStateOf(listOf<Product>()) }
 
@@ -112,7 +214,7 @@ class HomeUser : ComponentActivity() {
                                     .padding(bottom = 80.dp),
                                 color = MaterialTheme.colorScheme.background
                             ) {
-                                // Implement Maps 
+                                LocationScreen(this@HomeUser, currentLocation, cameraPositionState)
                             }
                         }
                         composable(BottomNavItem.Phone.route) {
@@ -129,6 +231,60 @@ class HomeUser : ComponentActivity() {
                     }
                 }
             )
+        }
+    }
+
+    @Composable
+    fun LocationScreen(context: Context, currentLocation: LatLng, camerapositionState: CameraPositionState) {
+
+        val launchMultiplePermissions = rememberLauncherForActivityResult(
+            ActivityResultContracts.RequestMultiplePermissions()) {
+                permissionMaps ->
+            val areGranted = permissionMaps.values.reduce {acc, next -> acc && next}
+            if (areGranted) {
+                locationRequired = true
+                startLocationUpdates()
+                Toast.makeText(context, "Permission Granted", Toast.LENGTH_SHORT).show()
+            }
+            else {
+                Toast.makeText(context, "Permission Denied", Toast.LENGTH_SHORT).show()
+            }
+        }
+
+        Box(modifier = Modifier.fillMaxSize()) {
+            GoogleMap (
+                modifier = Modifier.fillMaxSize(),
+                cameraPositionState = camerapositionState
+            ) {
+                Marker(
+                    state = MarkerState(
+                        position = currentLocation,
+                    ),
+                    title = "You",
+                    snippet = "You're here!!!"
+                )
+            }
+            Column(
+                modifier = Modifier.fillMaxSize(),
+                verticalArrangement = Arrangement.Bottom,
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                Text(text = "Your Location: ${currentLocation.latitude}/${currentLocation.longitude}")
+                Button(onClick = {
+                    if (permissions.all {
+                            ContextCompat.checkSelfPermission(context, it) == PackageManager.PERMISSION_GRANTED
+                        })
+                    {
+                        // Get Location
+                        startLocationUpdates()
+                    }
+                    else {
+                        launchMultiplePermissions.launch(permissions)
+                    }
+                }) {
+                    Text(text = "Get Your Location")
+                }
+            }
         }
     }
 }
